test(e2e): migrate trending smoke tests from Detox to Appium
Migrate trending search, feed, and browser smoke specs to
tests/smoke-appium/trending/ following the existing Appium smoke
pattern (appiumTest, loginToAppPlaywright, currentDeviceDetails).

Remove the Detox counterparts from tests/smoke/trending/ and register
SmokeWalletPlatform in Appium smoke CI workflows for iOS and Android.

MMQA-2005
